# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unify - Student Networking and Collaboration Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the authenticated user's profile (creates one on first call)
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfile"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update the authenticated user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfile"
  /me/onboarding:
    post:
      operationId: completeOnboarding
      tags: [onboarding]
      summary: Submit onboarding answers (username, interests, skills, bio)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfile"
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get public profile for a user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
...
components:
  schemas:
...
    Interest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        emoji:
          type: ["string", "null"]
      required: [id, name, category]
...
    MyProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin, moderator, test]
        coins:
          type: integer
        avatarColor:
          type: string
        bannerColor:
          type: string
        mobileNumber:
          type: ["string", "null"]
        usn:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        branch:
          type: ["string", "null"]
        yearEnrolled:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        interests:
          type: array
          items:
            $ref: "#/components/schemas/Interest"
        skills:
          type: array
          items:
            type: string
        communityCount:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
      required:
        - id
        - role
        - coins
        - avatarColor
        - bannerColor
        - onboardingComplete
        - interests
        - skills
        - communityCount
        - followerCount
        - followingCount
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        bio:
          type: ["string", "null"]
        avatarColor:
          type: string
        coins:
          type: integer
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        followerCount:
          type: integer
        followingCount:
          type: integer
        communityCount:
          type: integer
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - avatarColor
        - coins
        - skills
        - interests
        - followerCount
        - followingCount
        - communityCount
        - isFollowing
...
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        interestIds:
          type: array
          items:
            type: integer
        avatarColor:
          type: string
        bannerColor:
          type: string
        mobileNumber:
          type: string
        usn:
          type: string
        name:
          type: string
        semester:
          type: string
        branch:
          type: string
        yearEnrolled:
          type: string
        portfolioUrl:
          type: string
        linkedinUrl:
          type: string
    OnboardingBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 24
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        interestIds:
          type: array
          items:
            type: integer
      required: [username, interestIds]
...